agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Implementation of agent"
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "datarobot-genai[langgraph, dragent]==0.8.8",
    "datarobot-moderations==11.2.20",
    "datarobot-drum==1.17.11.post1",  # Lift upper limit after asyncio migration finished
    "datarobot-mlops==11.1.0",
    "openai>=1.81.0,<2.0.0",
    "litellm>=1.72.1,<2.0.0,!=1.82.7,!=1.82.8",
    "langchain-litellm>=0.2.3",
    "opentelemetry-api>=1.33.0,<2.0.0",
    "opentelemetry-sdk>=1.33.0,<2.0.0",
    "pdfminer.six>=20251230",  # CVE fix
    "langgraph-checkpoint>=3.0.0", # CVE fix
    "nltk>=3.9.3", # CVE fix
    "anyio>=4.11.0,<4.12.0", # Fix: ImportError: `set_current_async_library`
    "packaging>=25.0,<26.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.6.1",
    "pytest-cov>=6.1.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=1.2.0",
    "requests-mock>=1.12.1",
    "ruff>=0.12.0",
    "types-requests>=2.32.4.20250913",
    "yamlfix>=1.17.0",
    "debugpy>=1.8.20",
]
# Additional dependencies required to run the agent in DataRobot Agentic playground
agentic_playground = [
    "fastapi[All]==0.118.2",
    "psutil==5.9.8",
    "ipykernel<6.29.0",
    "jupyter-client>=8.6.3",
    "jupyter-core>=5.8.1",
    "jupyter-kernel-gateway>=3.0.1",
    "ecs-logging>=2.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pycache__",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "docker_context",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.yamlfix]
allow_duplicate_keys = true
line_length = 120

[tool.mypy]
strict = true
python_version = "3.11"
plugins = "pydantic.mypy"
exclude = """(?x)(
    output/ # exclude files starting with "output"
    | deployment_.*/
    | docker_context/
    | include/   # or files starting with "include."
    | venv/
    | .venv/
    | .cache
    | e2e/
    | tests/
    | build/
  )"""
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
disable_error_code = ["typeddict-unknown-key"]
enable_error_code = "ignore-without-code"
implicit_reexport = true
ignore_missing_imports = true
no_implicit_optional = true
show_column_numbers = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "datarobot.*"
implicit_reexport = true
follow_imports = "skip"

[tool.uv]
required-version = ">=0.7.0"
override-dependencies = [
    "fastapi>=0.118.2,<0.133.0",  # 0.132 is the last version compatiable with NAT 1.4.1
    "ragas>=0.3.8",
    "starlette>=0.49.1,<1.0.0",  # 0.52.1 is the last version compatiable with NAT 1.4.1
    "pyarrow==21.0.0",
    # There is an incompatiblity between the opentelemetry-semantic-conventions-ai==0.4.14
    # and crewai==1.9.3. See BUZZOK-29568 for more info.
    "opentelemetry-semantic-conventions-ai==0.4.13",
    # nvidia-nat-crewai==1.4.1 pins crewai<1.0.0 but datarobot-genai requires crewai>=1.1.0.
    # Override until nvidia-nat-crewai ships a fix upstream (BUZZOK-29844).
    # Upstream issue: https://github.com/NVIDIA/NeMo-Agent-Toolkit/issues/1802
    "crewai[litellm]>=1.1.0,<2.0.0",
    "litellm>=1.72.1,<2.0.0,!=1.82.7,!=1.82.8",
    "gevent; sys_platform == 'never'",  # not used; therefore removed
    "onnxruntime; sys_platform == 'never'",  # not used; therefore removed
    "fastembed; sys_platform == 'never'",  # not used; therefore removed
]
package = true
environments = [
  "sys_platform == 'darwin'",
  "sys_platform == 'linux'",
  "sys_platform == 'win32'"
]

# This can be set to allow for proxy indexes in air-gapped environments
# [[tool.uv.index]]
# name = "mirror"
# url = "https://mirror_url/"

[dependency-groups]
extras = []


[project.entry-points.'nat.plugins']
langgraph_agent = "agent.register"


[tool.setuptools.packages.find]
include = ["agent*"]
